requirements.txt
Flask==3.0.1
Flask-Cors==5.0.0
Flask-Migrate==2.5.2
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
Redis==4.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==1.4.46